API

Main analysis functions

TransitionsInTimeseries.TransitionsSurrogatesConfig — Type

TransitionsSurrogatesConfig(indicators, change_metrics [, surrogate]; kwargs...)

A configuration struct for TransitionsInTimeseries.jl that can be given to estimate_transitions. It contains all information necessary to perform the basic Workflow of the package to detect significant transitions in the input timeseries, using the method of surrogate testing to quantify significance:

Estimate the timeseries of an indicator by sliding a window over the input timeseries.

Estimate changes for an indicator by sliding a window over its timeseries.

Generate many surrogates that preserve important statistical properties of the original input timeseries.

Perform steps 1 and 2 for the surrogate timeseries (and for all provided indicators).

Estimate when an indicators timeseries shows significant change (trend, jump or anything else) when compared to the surrogate timeseries when compared to the surrogate data. Significance is estimated from the p-values of the real data vs surrogate data.

Arguments

indicators::AbstractVector{<:Function} a vector of indicators. Some indicators typically used in the literature are listed in the documentation section on indicators. The analysis is performed efficiently for all indicators given.

change_metrics change metrics corresponding to the given indicators. If given a function, the same function is used for all indicators. Otherwise, it should be a vetor of functions of the same size as indicators, using change metric for its corresponding indicator. Some change metrics typically used in the literature are listed in the documentation section on change metrics.

surrogate::Surrogate the method to use to generate surrogates of the input timeseries. This is an optional argument that defaults to RandomFourier(), see Surrogates for alternative options.

Both indicators and change metrics are generic Julia functions that input an x::AbstractVector and output an s::Real. Any appropriate function may be given and see making custom indicators/change metrics in the documentation for more information.

Keyword arguments

width_ind::Int=100, stride_ind::Int=1: width and stride given to the WindowViewer to compute the indicator from the input timeseries.

width_cha::Int=50, stride_cha::Int=1: width and stride given to the WindowViewer to compute the change metric timeseries from the indicator timeseries.

whichtime = midpoint: The time vector corresponding to the indicators / change metric timeseriesm is obtained from t using the keyword whichtime. Options include:

last: use the last timepoint of each window

midpoint: use the mid timepoint of each time window

first: use first timepoint of each window

In fact, the indicators time vector is computed simply via

t_indicator = windowmap(whichtime, t; width_ind, stride_ind)
t_change = windowmap(whichtime, t_indicator; width_cha, stride_cha)

so any other function of the time window may be given to extract the time point itself, such as mean or median.

n_surrogates::Int = 10_000: how many surrogates to create.

rng::AbstractRNG = Random.default_rng()`: a random number generator for the surrogates.

tail::Symbol = :both: kind of tail test to do (one of :left, :right, :both) when estimating the p-value from the distribution of surrogate data.

T = Float64: Element type of input timeseries to initialize some computations.

TransitionsInTimeseries.estimate_transitions — Function

estimate_transitions([t, ] x, config::TransitionsSurrogatesConfig)

Estimate possible transitions for input timeseries x according to the configuration. If t (the time vector of x), is not provided, it is assumed t = eachindex(x).

Return the output as TransitionsResults. You can use this output also in transition_flags.

Missing docstring.

Missing docstring for TransitionsResults. Check Documenter's build log for details.

TransitionsInTimeseries.transition_flags — Function

transition_flags(results::TransitionsResults, p_threshold::Real) → flags

Return flags::Matrix{Bool}, which is an n × (c+1) sized matrix. Each column flags[:, c] is the Boolean flags that correspond to a p-value below the threshold p for each change metric timeseries. There are c change metric timeseries in total, but the c+1-th column of flags is simply the Boolean addition of all previous columns (i.e., time points where all indicators pass the significance).

To obtain the time windows where the c-th change metric is significant, simply do:

timepoints = results.t_change[flags[:, c]]

Indicators

Value distribution

Statistics.mean — Method

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

StatsBase.skewness — Function

skewness(v, [wv::AbstractWeights], m=mean(v))

Compute the standardized skewness of a real-valued array v, optionally specifying a weighting vector wv and a center m.

StatsBase.kurtosis — Function

kurtosis(v, [wv::AbstractWeights], m=mean(v))

Compute the excess kurtosis of a real-valued array v, optionally specifying a weighting vector wv and a center m.

Critical Slowing Down

Statistics.var — Method

var(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample variance of collection itr.

The algorithm returns an estimator of the generative distribution's variance under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sum((itr .- mean(itr)).^2) / (length(itr) - 1). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false where n is the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the variance over dimensions.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the variance of non-missing values.

TransitionsInTimeseries.ar1_whitenoise — Function

ar1_whitenoise(x::AbstractVector)

Return the AR1 regression coefficient θ of a time series x. Computation based on the analytic solution of the least-square parameter estimation:

\[\theta = \sum_{i=2}^{n} x_i x_{i-1} / \sum_{i=2}^{n} x_{i-1}^2\]

Spectrum

TransitionsInTimeseries.LowfreqPowerSpectrum — Type

LowfreqPowerSpectrum(; q_lofreq = 0.1)

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedLowfreqPowerSpectrum. The latter can be initialized by precompute:

lfps = precompute( LowfreqPowerSpectrum() )

Keyword arguments:

q_lofreq: a number between 0 and 1 that characterises which portion of the

frequency spectrum is considered to be low. For instance, q_lofreq = 0.1 implies that the lowest 10% of frequencies are considered to be the low ones.

TransitionsInTimeseries.PrecomputedLowfreqPowerSpectrum — Type

PrecomputedLowfreqPowerSpectrum(x::AbstractVector)

A struct containing all the precomputed fields to efficiently perform repetitive computation of the low-frequency power spectrum (LFPS), a number between 0 and 1 that characterizes the amount of power contained in the low frequencies of the power density spectrum of x. Once lfps::PrecomputedLowfreqPowerSpectrum is initialized, it can be used as a function to obtain the LFPS of x::AbstractVector by:

lfps(x)

Nonlinear dynamics

Indicators that come from nonlinear timeseries analysis and quantify some entropy-based dynamic quantity in the timeseries. They are provided by the ComplexityMeasures.jl package, that lists 100s of possible such indicators. Here we only provide an indicator out of the box for the permutation entropy, but building something similar is trivial:

function PermutationEntropy(; m = 3, τ = 1)
    est = SymbolicPermutation(; m, τ)
    return x -> entropy_normalized(est, x)
end

TransitionsInTimeseries.PermutationEntropy — Function

PermutationEntropy(; m = 3, τ = 1) → f

Return a function that given timeseries x it computes the normalized permutation entropy of order m with time delay τ.

Missing docstring.

Missing docstring for entropy. Check Documenter's build log for details.

Change metrics

Slope

TransitionsInTimeseries.kendalltau — Function

kendalltau(x::AbstractVector)

Compute the kendall-τ correlation coefficient of the time series x.

TransitionsInTimeseries.spearman — Function

spearman(x::AbstractVector)

Compute the spearman correlation coefficient of the time series x.

TransitionsInTimeseries.RidgeRegressionSlope — Type

RidgeRegressionSlope(; lambda = 0.0)

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedRidgeRegressionSlope. The latter can be initialized by precompute:

rr = precompute(RidgeRegressionSlope(), x)

Keyword arguments:

lambda: a regularization constant, usually between 0 and 1.

TransitionsInTimeseries.PrecomputedRidgeRegressionSlope — Type

PrecomputedRidgeRegressionSlope(x::AbstractVector)

Return the slope of the ridge regression of x.

Difference in value distribution

Make your own indicator/metric!

The only difference between what is an "indicator" and what is a "change metric" is purely conceptual. As far as the code base of TransitionsInTimeseries.jl is concerned, they are both functions f: x::AbstractVector{Real} -> f(x)::Real. As a user you may give any such function for an indicator or change metric.

There are situations where you may optimize such a function based on knowledge of input x type and length.

TODO: Here explain how to use precomputable functions

TransitionsInTimeseries.PrecomputableFunction — Type

PrecomputableFunction

Supertype of structs containing the necessary field to precompute a ::Function by:

precompute(f::PrecomputableFunction, t)

Missing docstring.

Missing docstring for precompute. Check Documenter's build log for details.

Surrogates

For the surrogate generation, you can use any subtype of Surrogate defined in Timeseriessurrogates.jl.

Sliding windows

TransitionsInTimeseries.WindowViewer — Type

WindowViewer(x; width, stride)

Initialize an iterator that generates views over the given timeseries x based on a window with a given width, incrementing the window views with the given stride. You can use this directly with map, such as map(std, WindowViewer(x, ...)) would give you the moving-window-timeseries of the std of x.

If not given, the keywords width, stride are taken as default_window_width(x) and 1.

TransitionsInTimeseries.windowmap — Function

windowmap(f::Function, x::AbstractVector; kwargs...) → mapped_f

A shortcut for first generating a wv = WindowViewer(x; kwargs...) and then applying mapped_f = map(f, wv). If x is accompanied by a time vector t, you probably also want to call this function with t instead of x and with one of mean, midpoint, midvalue as f to obtain a time vector for the mapped_f output.

TransitionsInTimeseries.windowmap! — Function

windowmap!(f::Function, out, x::AbstractVector; kwargs...)

Same as windowmap, but writes the output in-place in out.

Load data

TransitionsInTimeseries.load_linear_vs_doublewell — Method

load_linear_vs_doublewell()

Load prototypical data from a linear and a double-well model to test some indicators.