TransitionIndicators.jl

TransitionIndicators — Module

TransitionIndicators.jl

A Julia package that can estimates indicators of transitions (from one dynamic regime or stable state to another) in timeseries. Also bundles the indicators with significance testing via surrogate analysis using TimeseriesSurrogates.jl. Alternative names for this package could have been: Early Warning Signals / Resilience Indicators / Regime Shift Identifiers / Change Point Detectors, or however else you want to call them!

To install it, run import Pkg; Pkg.add("TransitionIndicators").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Example

using TransitionIndicators
using Statistics: mean, var

n = 1001
t = collect(1.0:n)
x = copy(t)
p = init_metaanalysis_params(n_surrogates = 100)
res = analyze_indicators(t, x, [mean, var], ridge_slope, p)

# The trend of mean(windowview) is the stride for x=t
meantrend_ground_truth = fill(p.wv_indicator_stride, length(res.t_evolution))
# The trend of var(windowview) is 0 for x any affine function of t.
vartrend_ground_truth = fill(0.0, length(res.t_evolution))

201-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 ⋮
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0

API

TransitionIndicators.analyze_indicators — Function

analyze_indicators(t, x, indicators, evolution_metrics, p)

Compute the indicators and their evolution_metrics for a timeseries t, x and its surrogates. If t is not provided, it is simply assumed to be 1:length(x). This meta-analysis is performed based on p::MetaAnalysisParameters.