API

Load data

TransitionIndicators.load_linear_vs_doublewell — Method

load_linear_vs_doublewell()

Load prototypical data from a linear and a double-well model to test some indicators.

Indicators analysis

TransitionIndicators.IndicatorsConfig — Type

IndicatorsConfig(indicators...; window_kwargs...)

A configuration for computing indicators from timeseries. Indicators are standard Julia functions that input an AbstractVector and output a real number. Any number of indicators can be given. Indicators typically used in the literature are listed in the documentation section Indicators.

Keywords are propagated into WindowViewer to create a sliding window for estimating the indicators.

Along with SignificanceConfig it is given to indicators_analysis.

TransitionIndicators.SignificanceConfig — Type

SignificanceConfig(change_metrics...; kwargs...)

A configuration for estimating a significant change of indicators computed from timeseries. Along with IndicatorsConfig it is given to indicators_analysis.

change_metrics can be a single Julia function, in which case the same metric is applied over all indicators in IndicatorsConfig. change_metrics can also be many functions, each one for each indicator.

Keyword arguments

n_surrogates::Int = 10_000: how many surrogates to create.

surrogate_method::S = RandomFourier(): what method to use to create the surrogates. Any Surrogate subtype from TimeseriesSurrogates.jl is valid.

rng::AbstractRNG = Random.default_rng()`: a random number generator for the surrogates. See the Julia manual for more.

wv_indicator_width::Int = 100,

wv_indicator_stride::Int = 5,

width, stride = 20, 1: width and stride given to the WindowViewer of the indicator timeseries. Notice that here the default values are different.

TransitionIndicators.indicators_analysis — Function

indicators_analysis(x, indicators::IndicatorsConfig, significance::SignificanceConfig)

Perform an analysis of transition indicators for input timeseries x by specifying which indicators to use and with what sliding window (IndicatorsConfig), as well as how to measure significant changes in the indicators (SignificanceConfig).

Return the output as IndicatorsResults.

This function performs the analysis described in the documentation Example sections. It computes various indicators over sliding windows of x, and then computes change metrics of over sliding windows of the indicators. In parallel it does exactly the same computations for surrogates of x. The returned output contains all these computed timeseries and can be given to indicators_significance.

Indicators significance

TransitionIndicators.indicators_significance — Function

indicators_significance(res::IndicatorsResults, q::SignificanceTest) → out

Given the output of indicators_analysis, estimate when the computed indicators have a statistically significant change, with "significant" defined by q, which is a subtype of SignificanceTest.

The output is returned as an n×m matrix with n the length of the time vector of the change metric timeseries, and m the number of change metrics (which is the same as the number of indicators). The values of the matrix are Booleans (true for significant).

You can use prod(out; dim = 2) for logical & product across indicators.

Missing docstring.

Missing docstring for SignificanceTest. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Quantile. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Sigma. Check Documenter's build log for details.

Missing docstring.

Missing docstring for significant. Check Documenter's build log for details.

Indicators

TransitionIndicators.ar1_whitenoise — Function

ar1_whitenoise(x) → ̂θ

Estimate the AR1 regression coefficient θ of a vector time series x. Computation based on the analytic solution of the least-square parameter estimation:

\[ \hat{\theta} = \dfrac{\sum_{i=2}^{n} x_i \, x_{i-1}}{\sum_{i=2}^{n} x_{i-1}^2}\]

Statistics.mean — Method

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

Statistics.var — Method

var(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample variance of collection itr.

The algorithm returns an estimator of the generative distribution's variance under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sum((itr .- mean(itr)).^2) / (length(itr) - 1)). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false where n is the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the variance over dimensions.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the variance of non-missing values.

StatsBase.skewness — Function

skewness(v, [wv::AbstractWeights], m=mean(v))

Compute the standardized skewness of a real-valued array v, optionally specifying a weighting vector wv and a center m.

StatsBase.kurtosis — Function

kurtosis(v, [wv::AbstractWeights], m=mean(v))

Compute the excess kurtosis of a real-valued array v, optionally specifying a weighting vector wv and a center m.

Change metrics

TransitionIndicators.RidgeRegression — Type

RidgeRegression(t, width; lambda = 0.0) → rr

Initialize a ridge regression for a time vector t, a sliding-window width and an optional regularizeation term lambda. The output rr can then be used as a function to perform the regression and extract the slope! If lambda = 0, linear regression is recovered (default case). For more information, visit: https://en.wikipedia.org/wiki/Ridge_regression.

Examples

julia> t = 0.0:1.0:100;
julia> x = 2.3 .* t .+ 1.2;
julia> rr = RidgeRegression(t);
julia> rr(x)
2.299999999999999

TransitionIndicators.kendalltau — Function

kendalltau(x)

Compute the kendall-τ correlation coefficient of the time series x.

TransitionIndicators.spearman — Function

spearman(x)

Compute the spearman correlation coefficient of the time series x.

Surrogates

For the surrogate generation, you can use any subtype of Surrogate defined in Timeseriessurrogates.jl.

Significance metrics

Missing docstring.

Missing docstring for threshold_indicators. Check Documenter's build log for details.

Missing docstring.

Missing docstring for measure_significance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for confidence_interval. Check Documenter's build log for details.

Missing docstring.

Missing docstring for normalized_percentile. Check Documenter's build log for details.

Sliding windows

TransitionIndicators.WindowViewer — Type

WindowViewer(x; width, stride)

Initialize an iterator that generates views over the given timeseries x based on a window with a given width, incrementing the window views with the given stride. You can use this directly with map, such as map(std, WindowViewer(x, ...)) would give you the moving-window-timeseries of the std of x.

If not given, the keywords width, stride as taken as length(x)÷100 and 1.

TransitionIndicators.windowmap — Function

windowmap(f::Function, x::AbstractVector; kwargs...) → mapped_f

A shortcut for first generating a wv = WindowViewer(x; kwargs...) and then applying mapped_f = map(f, wv). If x is accompanied by a time vector t, you probably also want to call this function with t instead of x and with one of mean, midpoint, midvalue as f to obtain a time vector for the mapped_f output.

TransitionIndicators.windowmap! — Function

windowmap!(f::Function, out, x::AbstractVector; kwargs...)

Same as windowmap, but writes the output in-place in out.