API

Main analysis functions

TransitionsInTimeseries.WindowedIndicatorConfig — Type

WindowedIndicatorConfig(indicators, change_metrics; kwargs...) → config

A configuration struct for TransitionsInTimeseries.jl that collects what indicators and corresponding metrics to use in the transitions_analysis.

indicators is a tuple of indicators (or a single indicator). change_metrics is also a tuple or a single function. If a single function, the same change metric is used for all provided indicators. This way the analysis can be efficiently repeated for many indicators and/or change metrics.

Both indicators and change metrics are generic Julia functions that input an x::AbstractVector and output an s::Real. Any appropriate function may be given and see making custom indicators/change metrics in the documentation for more information.

Keyword arguments

width_ind::Int=100, stride_ind::Int=1: width and stride given to WindowViewer to compute the indicator from the input timeseries.

width_cha::Int=50, stride_cha::Int=1: width and stride given to WindowViewer to compute the change metric timeseries from the indicator timeseries.

whichtime = midpoint: The time vector corresponding to the indicators / change metric timeseries is obtained from t in transitions_analysis using the keyword whichtime. Options include:

last: use the last timepoint of each window

midpoint: use the mid timepoint of each time window

first: use first timepoint of each window

In fact, the indicators time vector is computed simply via

t_indicator = windowmap(whichtime, t; width_ind, stride_ind)
t_change = windowmap(whichtime, t_indicator; width_cha, stride_cha)

so any other function of the time window may be given to extract the time point itself, such as mean or median.

T = Float64: Element type of input timeseries to initialize some computations.

TransitionsInTimeseries.transition_metrics — Function

transition_metrics(config::WindowedIndicatorConfig, x [,t]) → output

Estimate possible transitions for input timeseries x using a sliding window approach as described by config:

Estimate the timeseries of an indicator by sliding a window over the input timeseries.

Estimate changes of an indicator by sliding a window of the change metric over the indicator timeseries.

If t (the time vector of x), is not provided, it is assumed t = eachindex(x).

Return the output as WindowedIndicatorResults which can be given to estimate_significance! to deduce which possible transitions are statistically significant using a variety of significance tests.

TransitionsInTimeseries.WindowedIndicatorResults — Type

WindowedIndicatorResults

A struct containing the output of transition_metrics used with WindowedIndicatorConfig. It can be used for further analysis, visualization, or given to estimate_significance!.

It has the following fields that the user may access

x: the input timeseries.

t: the time vector of the input timeseries.

x_indicator, the indicator timeseries (matrix with each column one indicator).

t_indicator, the time vector of the indicator timeseries.

x_change, the change metric timeseries (matrix with each column one change metric).

t_change, the time vector of the change metric timeseries.

wim::WindowedIndicatorConfig, used for the analysis.

Significance testing

TransitionsInTimeseries.estimate_significance — Function

estimate_significance(sigconfig::SignificanceConfig, res::WindowedIndicatorResults)

Estimate the significance of the transition metrics stored in res by using the method described by sigconfig and return the results in TransitionsSignificance.

TransitionsInTimeseries.SignificanceConfig — Type

SignificanceConfig

Supertype used to test for significance in estimate_significance. Valid subtypes are:

SurrogatesConfig.

TransitionsInTimeseries.SurrogatesConfig — Type

SurrogatesConfig <: SignificanceConfig
SurrogatesConfig(; surrogate = RandomFourier(), n = 1_000, tail = :both, rng)

A configuration struct for significance testing estimate_significance using timeseries surrogates.

Keyword arguments

surromethod = RandomFourier(): method to generate surrogates

n = 1_000: how many surrogates to generate

rng = Random.default_rng(): random number generator for the surrogates

p = 0.05: threshold for significance of the p-value

tail = :both: tail type used, see below

detrend_surro = false: Boolean specifying whether the surrogates are to be

detrended by a linear regression. This is sometimes performed in the literature but seldomly impacts the results in a significant way, since the surrogates of the residual should already be largely free of trend.

Description

When used with WindowedIndicatorResults, significance is estimated as follows: n surrogates from the input timeseries are generated using surromethod, which is any Surrogate subtype provided by TimeseriesSurrogates.jl. For each surrogate, the indicator and then change metric timeseries is extracted. The values of the surrogate change metrics form a distribution of values (one at each time point). The value of the original change metric is compared to that of the surrogate distribution and a p-value is extracted according to the specified tail. The p-value is compared with p to claim significance, which results in flags. These results are then stored in SurrogatesSignificance.

The p-value is simply the proportion of surrogate change metric values that exceed (for tail = :right) or subseed (tail = :left) the original change metric at each given time point. Use tail = :left if the surrogate data are expected to have higher change metric, discriminatory statistic values. This is the case for statistics that quantify entropy. For statistics that quantify autocorrelation, use tail = :right instead. For anything else, use the default tail = :both. An iterable of tail values can also be given, in which case a specific tail is used for each change metric in WindowedIndicatorResults.

TransitionsInTimeseries.TransitionsSignificance — Type

TransitionsSignificance

Supertype used to store the results of estimate_significance based on a SignificanceConfig. Valid subtypes are:

SurrogatesSignificance.

TransitionsInTimeseries.SurrogatesSignificance — Type

SurrogatesSignificance

A struct containing the results of estimate_significance based on a SurrogatesConfig, i.e. the p-values and their associated flags. The former gives a continuous value between 0 and 1, with 0 denoting an extremely unlikely and therefore probably significant behavior of the original time series compared to the ensemble of surrogates. The latter is a Boolean matrix which contains true wherever the p-values are below the threshold defined in SurrogatesConfig. Both signif.p_values and signif.flags have the same size as res.x_change, with signif::SurrogatesSignificance and res::WindowedIndicatorResults.

TransitionsInTimeseries.QuantileSignificance — Type

QuantileSignificance(; p = 0.95, tail = :right) <: TransitionsSignificance

A configuration struct for significance testing significant_transitions. When used with WindowedIndicatorResults, significance is estimated as by comparing the value of each change metric with its p-quantile. Values that exceed the p-quantile (if tail = :right) or subseed the 1-p-quantile (if tail = :left) are deemed significant. If tail = :both then either condition is checked.

Indicators

Value distribution

Statistics.mean — Method

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

StatsBase.skewness — Function

skewness(v, [wv::AbstractWeights], m=mean(v))

Compute the standardized skewness of a real-valued array v, optionally specifying a weighting vector wv and a center m.

StatsBase.kurtosis — Function

kurtosis(v, [wv::AbstractWeights], m=mean(v))

Compute the excess kurtosis of a real-valued array v, optionally specifying a weighting vector wv and a center m.

Critical Slowing Down

Statistics.var — Method

var(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample variance of collection itr.

The algorithm returns an estimator of the generative distribution's variance under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sum((itr .- mean(itr)).^2) / (length(itr) - 1). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false where n is the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the variance over dimensions.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the variance of non-missing values.

TransitionsInTimeseries.ar1_whitenoise — Function

ar1_whitenoise(x::AbstractVector)

Return the AR1 regression coefficient $θ$ of a time series $x$ by computing the analytic solution of the least-square parameter estimation under white-noise assumption for the data-generating process:

\[\theta = \sum_{i=2}^{n} x_i x_{i-1} / \sum_{i=2}^{n} x_{i-1}^2\]

Spectrum

TransitionsInTimeseries.LowfreqPowerSpectrum — Type

LowfreqPowerSpectrum(; q_lofreq = 0.1)

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedLowfreqPowerSpectrum. The latter can be initialized by precompute:

lfps = precompute( LowfreqPowerSpectrum() )

Keyword arguments:

q_lofreq: a number between 0 and 1 that characterises which portion of the

frequency spectrum is considered to be low. For instance, q_lofreq = 0.1 implies that the lowest 10% of frequencies are considered to be the low ones.

Nonlinear dynamics

Indicators that come from nonlinear timeseries analysis and quantify some entropy-based dynamic quantity in the timeseries. They are provided by the ComplexityMeasures.jl package, that lists 100s of possible such indicators. Here we only provide an indicator out of the box for the permutation entropy, but building something similar is trivial:

function permutation_entropy(; m = 3, τ = 1)
    est = SymbolicPermutation(; m, τ)
    return x -> entropy_normalized(est, x)
end

TransitionsInTimeseries.permutation_entropy — Function

permutation_entropy(; m = 3, τ = 1) → f

Return a function that given timeseries x it computes the normalized permutation entropy of order m with time delay τ.

StatsBase.entropy — Function

entropy(p, [b])

Compute the entropy of a collection of probabilities p, optionally specifying a real number b such that the entropy is scaled by 1/log(b). Elements with probability 0 or 1 add 0 to the entropy.

Change metrics

Slope

TransitionsInTimeseries.kendalltau — Function

kendalltau(x::AbstractVector)

Compute the kendall-τ correlation coefficient of the time series x. kendalltau can be used as a change metric focused on trend.

TransitionsInTimeseries.spearman — Function

spearman(x::AbstractVector)

Compute the spearman correlation coefficient of the time series x. spearman can be used as a change metric focused on trend.

TransitionsInTimeseries.RidgeRegressionSlope — Type

RidgeRegressionSlope(; lambda = 0.0) → rr

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedRidgeRegressionSlope. rr can be used as a change metric focused on trend.

lambda is a regularization constant, usually between 0 and 1.

Value distribution differences

TransitionsInTimeseries.difference_of_means — Function

difference_of_means(x::AbstractArray)

Return the absolute difference of the means of the first and second halfs of x. difference_of_means can be used as a change metric focused on value differences. Creating similar statistical differences using other moments instead of mean is trivial. In fact, the source of difference_of_means is just:

# assumes 1-based indexing
n = length(x)
x1 = view(x, 1:n÷2)
x2 = view(x, (n÷2 + 1):n)
return abs(mean(x1) - mean(x2))

difference_of_means can also sensibly be used for windows of size 2, in which case the change metric timeseries is the same as the abs.(diff(...)) of the indicator timeseries.

Make your own indicator/metric!

The only difference between what is an "indicator" and what is a "change metric" is purely conceptual. As far as the code base of TransitionsInTimeseries.jl is concerned, they are both functions f: x::AbstractVector{Real} -> f(x)::Real. As a user you may give any such function for an indicator or change metric.

There are situations where you may optimize such a function based on knowledge of input x type and length.

TODO: Here explain how to use precomputable functions

TransitionsInTimeseries.PrecomputableFunction — Type

PrecomputableFunction

Supertype of structs containing the necessary field to precompute a ::Function by:

precompute(f::PrecomputableFunction, t)

Missing docstring.

Missing docstring for precompute. Check Documenter's build log for details.

Surrogates

For the surrogate generation, you can use any subtype of Surrogate defined in Timeseriessurrogates.jl.

Sliding windows

TransitionsInTimeseries.WindowViewer — Type

WindowViewer(x; width, stride)

Initialize an iterator that generates views over the given timeseries x based on a window with a given width, incrementing the window views with the given stride. You can use this directly with map, such as map(std, WindowViewer(x, ...)) would give you the moving-window-timeseries of the std of x.

If not given, the keywords width, stride are taken as default_window_width(x) and 1.

TransitionsInTimeseries.windowmap — Function

windowmap(f::Function, x::AbstractVector; kwargs...) → mapped_f

A shortcut for first generating a wv = WindowViewer(x; kwargs...) and then applying mapped_f = map(f, wv). If x is accompanied by a time vector t, you probably also want to call this function with t instead of x and with one of mean, midpoint, midvalue as f to obtain a time vector for the mapped_f output.

TransitionsInTimeseries.windowmap! — Function

windowmap!(f::Function, out, x::AbstractVector; kwargs...)

Same as windowmap, but writes the output in-place in out.

Load data

TransitionsInTimeseries.load_linear_vs_doublewell — Method

load_linear_vs_doublewell()

Load prototypical data from a linear and a double-well model to test some indicators.