API

Load data

TransitionIndicators.load_linear_vs_doublewell — Method

load_linear_vs_doublewell()

Load prototypical data from a linear and a double-well model to test some indicators.

Indicators analysis

TransitionIndicators.PrecomputableFunction — Type

PrecomputableFunction

Supertype of structs containing the necessary field to precompute a ::Function by:

precompute(f::PrecomputableFunction, t)

TransitionIndicators.precompute_metrics — Function

precompute_metrics(metrics::Vector, t::AbstractVector)

Precompute functions contained in metrics that can be precomputed based on the time vector t. Relies on abstract type PrecomputableFunction.

The output f::Vector{Function} contains the ready-to-use metrics.

TransitionIndicators.IndicatorsConfig — Type

IndicatorsConfig(t::AbstractVector, f::Function, indicators...; kwargs...)

A configuration for computing indicators from timeseries with time t. Indicators are standard Julia functions that input an ::AbstractVector and output a ::Real. Any number of indicators can be given. Indicators typically used in the literature are listed in the documentation section Indicators.

Keywords are propagated into WindowViewer to create a sliding window for estimating the indicators. The time vector resulting from the sliding window is obtained under the hood by:

t_indicator = windowmap(f, t, kwargs...)

Common choices for f are:

last: use x[k-width:k] to estimate the indicator at time step k.

midpoint: use x[k-width÷2:k+width÷2] to estimate the indicator at time step k.

first: use x[k:k+width] to estimate the indicator at time step k.

Along with SignificanceConfig it is given to indicators_analysis.

TransitionIndicators.SignificanceConfig — Type

SignificanceConfig(indconfig, f, change_metrics...; kwargs...)

A configuration for estimating a significant change of indicators based on a sliding window and and indconfig::IndicatorsConfig. The time vector resulting from the sliding window is obtained under the hood by:

t_change = windowmap(f, indconfig.t_indicator, kwargs...)

Common choices for f are:

last: use x[k-width:k] to estimate the indicator change at time step k.

midpoint: use x[k-width÷2:k+width÷2] to estimate the indicator change

at time step k.

first: use x[k:k+width] to estimate the indicator change at time step k.

Along with IndicatorsConfig, SignificanceConfig is given to indicators_analysis.

change_metrics can be a single Julia function, in which case the same metric is applied over all indicators in IndicatorsConfig. change_metrics can also be many functions, each one for each indicator.

Keyword arguments

n_surrogates::Int = 10_000: how many surrogates to create.

surrogate_method::S = RandomFourier(): what method to use to create the surrogates. Any Surrogate subtype from TimeseriesSurrogates.jl is valid.

rng::AbstractRNG = Random.default_rng()`: a random number generator for the surrogates. See the Julia manual for more.

width::Int: width given to the WindowViewer of the indicator timeseries to estimate transient changes of the indicator.

stride::Int: stride given to the WindowViewer of the indicator timeseries to estimate transient changes of the indicator.

tail::Symbol: kind of tail test to do (one of :left, :right, :both).

TransitionIndicators.indicators_analysis — Function

indicators_analysis(t, x, indicators::IndicatorsConfig, sigconfig::SignificanceConfig)

Perform an analysis of transition indicators for input timeseries x with time vector t by specifying which indicators to use and with what sliding window (IndicatorsConfig), as well as how to measure significant changes in the indicators (SignificanceConfig). If t is not provided, it is assumed that t=eachindex(x).

Return the output as IndicatorsResults.

This function performs the analysis described in the documentation Example sections. It computes various indicators over sliding windows of x, and then computes change metrics of over sliding windows of the indicators. It does exactly the same computations for surrogates of x and use this result to check for significance of the change metrics by computing the p-value. The returned output contains all these computed timeseries.

TransitionIndicators.IndicatorsResults — Type

IndicatorsResults

A struct containing the output of indicators_analysis, which is the main computational part of TransitionIndicators.jl. It can be used for analysis and visualization purposes.

It has the following fields:

x: the input timeseries.

t: the time vector of the input timeseries.

indicators::Vector{Function}: indicators used in the processing.

x_indicator, the indicator timeseries (matrix with each column one indicator).

t_indicator, the time vector of the indicator timeseries.

change_metrics::Vector{Function}: change metrics used in the processing.

x_change, the change metric timeseries (matrix with each column one change metric).

t_change, the time vector of the change metric timeseries.

pval, the p-value of the change metrics w.r.t. the surrogates. It is a 2-dimensional array, where first dimension = time, second dimension = change metric. I.e. pval[:, k] will give the time series of p-value of the k-th change metric.

Indicators

Statistics.mean — Method

mean(itr)

Compute the mean of all elements in a collection.

Note

If itr contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the mean of non-missing values.

Examples

julia> using Statistics

julia> mean(1:20)
10.5

julia> mean([1, missing, 3])
missing

julia> mean(skipmissing([1, missing, 3]))
2.0

Statistics.var — Method

var(itr; corrected::Bool=true, mean=nothing[, dims])

Compute the sample variance of collection itr.

The algorithm returns an estimator of the generative distribution's variance under the assumption that each entry of itr is a sample drawn from the same unknown distribution, with the samples uncorrelated. For arrays, this computation is equivalent to calculating sum((itr .- mean(itr)).^2) / (length(itr) - 1). If corrected is true, then the sum is scaled with n-1, whereas the sum is scaled with n if corrected is false where n is the number of elements in itr.

If itr is an AbstractArray, dims can be provided to compute the variance over dimensions.

A pre-computed mean may be provided. When dims is specified, mean must be an array with the same shape as mean(itr, dims=dims) (additional trailing singleton dimensions are allowed).

Note

If array contains NaN or missing values, the result is also NaN or missing (missing takes precedence if array contains both). Use the skipmissing function to omit missing entries and compute the variance of non-missing values.

StatsBase.skewness — Function

skewness(v, [wv::AbstractWeights], m=mean(v))

Compute the standardized skewness of a real-valued array v, optionally specifying a weighting vector wv and a center m.

StatsBase.kurtosis — Function

kurtosis(v, [wv::AbstractWeights], m=mean(v))

Compute the excess kurtosis of a real-valued array v, optionally specifying a weighting vector wv and a center m.

TransitionIndicators.ar1_whitenoise — Function

ar1_whitenoise(x::AbstractVector)

Return the AR1 regression coefficient θ of a time series x. Computation based on the analytic solution of the least-square parameter estimation:

\[ \hat{\theta} = \dfrac{\sum_{i=2}^{n} x_i \, x_{i-1}}{\sum_{i=2}^{n} x_{i-1}^2}\]

TransitionIndicators.LowfreqPowerSpectrum — Type

LowfreqPowerSpectrum(; q_lofreq = 0.1)

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedLowfreqPowerSpectrum. The latter can be initialized by precompute:

lfps = precompute( LowfreqPowerSpectrum() )

Keyword arguments:

q_lofreq: a number between 0 and 1 that characterises which portion of the

frequency spectrum is considered to be low. For instance, q_lofreq = 0.1 implies that the lowest 10% of frequencies are considered to be the low ones.

TransitionIndicators.PrecomputedLowfreqPowerSpectrum — Type

PrecomputedLowfreqPowerSpectrum(x::AbstractVector)

A struct containing all the precomputed fields to efficiently perform repetitive computation of the low-frequency power spectrum (LFPS), a number between 0 and 1 that characterizes the amount of power contained in the low frequencies of the power density spectrum of x. Once lfps::PrecomputedLowfreqPowerSpectrum is initialized, it can be used as a function to obtain the LFPS of x::AbstractVector by:

lfps(x)

Missing docstring.

Missing docstring for PermutationEntropy(). Check Documenter's build log for details.

TransitionIndicators.PermutationEntropy — Method

PermutationEntropy(x::AbstractVector)

Return the permutation entropy of x. This computation breaks down in computing the entropy_normalized of a SymbolicPermutation.

Change metrics

TransitionIndicators.kendalltau — Function

kendalltau(x::AbstractVector)

Compute the kendall-τ correlation coefficient of the time series x.

TransitionIndicators.spearman — Function

spearman(x::AbstractVector)

Compute the spearman correlation coefficient of the time series x.

TransitionIndicators.RidgeRegressionSlope — Type

RidgeRegressionSlope(; lambda = 0.0)

Return a PrecomputableFunction containing all the necessary fields to generate a PrecomputedRidgeRegressionSlope. The latter can be initialized by precompute:

rr = precompute( RidgeRegressionSlope() )

Keyword arguments:

lambda: a regularization constant, usually between 0 and 1.

TransitionIndicators.PrecomputedRidgeRegressionSlope — Type

PrecomputedRidgeRegressionSlope(x::AbstractVector)

Return the slope of the ridge regression of x.

Surrogates

For the surrogate generation, you can use any subtype of Surrogate defined in Timeseriessurrogates.jl.

Sliding windows

TransitionIndicators.WindowViewer — Type

WindowViewer(x; width, stride)

Initialize an iterator that generates views over the given timeseries x based on a window with a given width, incrementing the window views with the given stride. You can use this directly with map, such as map(std, WindowViewer(x, ...)) would give you the moving-window-timeseries of the std of x.

If not given, the keywords width, stride are taken as default_window_width(x) and 1.

TransitionIndicators.windowmap — Function

windowmap(f::Function, x::AbstractVector; kwargs...) → mapped_f

A shortcut for first generating a wv = WindowViewer(x; kwargs...) and then applying mapped_f = map(f, wv). If x is accompanied by a time vector t, you probably also want to call this function with t instead of x and with one of mean, midpoint, midvalue as f to obtain a time vector for the mapped_f output.

TransitionIndicators.windowmap! — Function

windowmap!(f::Function, out, x::AbstractVector; kwargs...)

Same as windowmap, but writes the output in-place in out.